Changes - Unification of EndPoint users in a single file
